Remove parser improvements from TODOs
The parser is already structured so that parsing will terminate with
trying to load a "primary" entity, and throw an "unexpected token"
error. There's room to make messages for particular situations - I added
a case for an unterminated string - but I don't think there's any
particular gap that will cause it to explode... just mediocre error
messages.